requirements.txt
pandas>=1.5
numpy>=1.23
yfinance>=0.2
ta>=0.10
scikit-learn>=1.1
xgboost>=1.6
requests>=2.31
pyarrow>=12.0
numba>=0.57
lightgbm>=4.0
imbalanced-learn>=0.11
google-api-python-client>=2.137
google-auth>=2.20
google-auth-httplib2>=0.2